pyproject.toml
[project]
name = "Sviewer"
version = "4.1.0"
description = "Medical image viewer"
authors = [
    { name="Jinfr"}
]
requires-python = ">=3.10"
dependencies = [
    "numpy==1.23.5",
    "opencv-python==4.10.0.84",
    "PySide6==6.10.2",
    "SimpleITK==2.4.1",
    "vtk==9.5.0",
    "onnxruntime==1.19.2",
    "PyYAML==6.0.2"
]
